fix: remove feature flag, some polish fixes
