+++ src/featureFlags/flags/featureFlags.ts
@@ -2,5 +2,4 @@ export enum FeatureFlag {
   traceJsonRpc = 'traceJsonRpc',
   permit2 = 'permit2',
   landingRedirect = 'landingRedirect',
-  aboutLandingPage = 'aboutLandingPage',
 }
