+++ src/pages/Landing/index.tsx
@@ -1,8 +1,13 @@
+import { Trans } from '@lingui/macro'
 import { Trace, TraceEvent } from '@uniswap/analytics'
 import { BrowserEvent, ElementName, EventName, PageName } from '@uniswap/analytics-events'
 import { BaseButton } from 'components/Button'
+import Card, { CardType } from 'pages/About/Card'
+import { MAIN_CARDS, MORE_CARDS } from 'pages/About/constants'
+import ProtocolBanner from 'pages/About/ProtocolBanner'
 import Swap from 'pages/Swap'
-import { useEffect } from 'react'
+import { useRef } from 'react'
+import { ArrowDownCircle } from 'react-feather'
 import { useLocation } from 'react-router-dom'
 import { Link as NativeLink } from 'react-router-dom'
 import { useIsDarkMode } from 'state/user/hooks'
@@ -10,30 +15,57 @@ import styled from 'styled-components/macro'
 import { BREAKPOINTS } from 'theme'
 import { Z_INDEX } from 'theme/zIndex'
 
-const PageContainer = styled.div`
+const PageContainer = styled.div<{ isDarkMode: boolean }>`
+  position: absolute;
+  top: 0;
+  padding: ${({ theme }) => theme.navHeight}px 0px 0px 0px;
   width: 100%;
-  position: relative;
   display: flex;
   flex-direction: column;
   align-items: center;
+  scroll-behavior: smooth;
+  overflow-x: hidden;
 
-  height: ${({ theme }) => `calc(100vh - ${theme.navHeight + theme.mobileBottomBarHeight}px)`};
-  @media screen and (min-width: ${({ theme }) => theme.breakpoint.md}px) {
-    height: ${({ theme }) => `calc(100vh - ${theme.navHeight}px)`};
-  }
+  background: ${({ isDarkMode }) =>
+    isDarkMode
+      ? 'linear-gradient(rgba(8, 10, 24, 0) 0%, rgb(8 10 24 / 100%) 45%)'
+      : 'linear-gradient(rgba(255, 255, 255, 0) 0%, rgb(255 255 255 /100%) 45%)'};
 `
 
 const Gradient = styled.div<{ isDarkMode: boolean }>`
   position: absolute;
+  display: flex;
+  align-items: center;
+  justify-content: center;
   top: 0;
   bottom: 0;
   width: 100%;
+  min-height: 550px;
   background: ${({ isDarkMode }) =>
     isDarkMode
       ? 'linear-gradient(rgba(8, 10, 24, 0) 0%, rgb(8 10 24 / 100%) 45%)'
       : 'linear-gradient(rgba(255, 255, 255, 0) 0%, rgb(255 255 255 /100%) 45%)'};
   z-index: ${Z_INDEX.dropdown};
   pointer-events: none;
+  height: ${({ theme }) => `calc(100vh - ${theme.mobileBottomBarHeight}px)`};
+  @media screen and (min-width: ${({ theme }) => theme.breakpoint.md}px) {
+    height: 100vh;
+  }
+`
+
+const GlowContainer = styled.div`
+  position: absolute;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  top: 0;
+  bottom: 0;
+  width: 100%;
+  overflow-y: hidden;
+  height: ${({ theme }) => `calc(100vh - ${theme.mobileBottomBarHeight}px)`};
+  @media screen and (min-width: ${({ theme }) => theme.breakpoint.md}px) {
+    height: 100vh;
+  }
 `
 
 const Glow = styled.div`
@@ -45,27 +77,23 @@ const Glow = styled.div`
   border-radius: 24px;
   max-width: 480px;
   width: 100%;
+  height: 100%;
 `
 
 const ContentContainer = styled.div<{ isDarkMode: boolean }>`
+  position: absolute;
   display: flex;
   flex-direction: column;
   align-items: center;
+  justify-content: end;
   width: 100%;
+  padding: 0 0 32px;
   max-width: min(720px, 90%);
-  position: sticky;
-  bottom: 0;
+  min-height: 500px;
   z-index: ${Z_INDEX.dropdown};
-  padding: 32px 0 80px;
   transition: ${({ theme }) => `${theme.transition.duration.medium} ${theme.transition.timing.ease} opacity`};
-
-  * {
-    pointer-events: auto;
-  }
-
-  @media screen and (min-width: ${BREAKPOINTS.md}px) {
-    padding: 64px 0;
-  }
+  height: ${({ theme }) => `calc(100vh - ${theme.navHeight + theme.mobileBottomBarHeight}px)`};
+  pointer-events: none;
 `
 
 const TitleText = styled.h1<{ isDarkMode: boolean }>`
@@ -120,7 +148,7 @@ const LandingButton = styled(BaseButton)`
 `
 
 const ButtonCTA = styled(LandingButton)`
-  background: linear-gradient(10deg, rgba(255, 0, 199, 1) 0%, rgba(255, 159, 251, 1) 100%);
+  background: linear-gradient(93.06deg, #ff00c7 2.66%, #ff9ffb 98.99%);
   border: none;
   color: ${({ theme }) => theme.white};
   transition: ${({ theme }) => `all ${theme.transition.duration.medium} ${theme.transition.timing.ease}`};
@@ -144,6 +172,79 @@ const ButtonCTAText = styled.p`
 const ActionsContainer = styled.span`
   max-width: 300px;
   width: 100%;
+  pointer-events: auto;
+`
+
+const LearnMoreContainer = styled.div`
+  align-items: center;
+  color: ${({ theme }) => theme.textTertiary};
+  cursor: pointer;
+  font-size: 20px;
+  font-weight: 600;
+  margin: 36px 0 0;
+  display: flex;
+  visibility: hidden;
+  pointer-events: auto;
+  @media screen and (min-width: ${BREAKPOINTS.sm}px) {
+    visibility: visible;
+  }
+
+  transition: ${({ theme }) => `${theme.transition.duration.medium} ${theme.transition.timing.ease} opacity`};
+
+  &:hover {
+    opacity: 0.6;
+  }
+`
+
+const LearnMoreArrow = styled(ArrowDownCircle)`
+  margin-left: 14px;
+  size: 20px;
+`
+
+const AboutContentContainer = styled.div<{ isDarkMode: boolean }>`
+  padding: 0 24px 5rem;
+  width: 100%;
+  background: ${({ isDarkMode }) =>
+    isDarkMode
+      ? 'linear-gradient(179.82deg, rgba(0, 0, 0, 0) 0.16%, #050026 99.85%)'
+      : 'linear-gradient(179.82deg, rgba(255, 255, 255, 0) 0.16%, #eaeaea 99.85%)'};
+  @media screen and (min-width: ${BREAKPOINTS.md}px) {
+    padding: 0 96px 5rem;
+  }
+`
+
+const CardGrid = styled.div<{ cols: number }>`
+  display: grid;
+  gap: 12px;
+  width: 100%;
+  padding: 24px 0 0;
+
+  grid-template-columns: 1fr;
+  @media screen and (min-width: ${BREAKPOINTS.sm}px) {
+    // At this screen size, we show up to 2 columns.
+    grid-template-columns: ${({ cols }) =>
+      Array.from(Array(cols === 2 ? 2 : 1))
+        .map(() => '1fr')
+        .join(' ')};
+    gap: 32px;
+  }
+
+  @media screen and (min-width: ${BREAKPOINTS.lg}px) {
+    // at this screen size, always show the max number of columns
+    grid-template-columns: ${({ cols }) =>
+      Array.from(Array(cols))
+        .map(() => '1fr')
+        .join(' ')};
+    gap: 32px;
+  }
+`
+
+const LandingSwapContainer = styled.div`
+  height: ${({ theme }) => `calc(100vh - ${theme.mobileBottomBarHeight}px)`};
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
 `
 
 const LandingSwap = styled(Swap)`
@@ -166,32 +267,31 @@ const Link = styled(NativeLink)`
 export default function Landing() {
   const isDarkMode = useIsDarkMode()
 
+  const cardsRef = useRef<HTMLDivElement>(null)
+
   const location = useLocation()
   const isOpen = location.pathname === '/'
-
-  useEffect(() => {
-    document.body.style.overflow = 'hidden'
-    return () => {
-      document.body.style.overflow = 'auto'
-    }
-  }, [])
 
   if (!isOpen) return null
 
   return (
     <Trace page={PageName.LANDING_PAGE} shouldLogImpression>
-      <PageContainer>
-        <TraceEvent
-          events={[BrowserEvent.onClick]}
-          name={EventName.ELEMENT_CLICKED}
-          element={ElementName.LANDING_PAGE_SWAP_ELEMENT}
-        >
-          <Link to="/swap">
-            <LandingSwap />
-          </Link>
-        </TraceEvent>
-        <Glow />
+      <PageContainer isDarkMode={isDarkMode}>
+        <LandingSwapContainer>
+          <TraceEvent
+            events={[BrowserEvent.onClick]}
+            name={EventName.ELEMENT_CLICKED}
+            element={ElementName.LANDING_PAGE_SWAP_ELEMENT}
+          >
+            <Link to="/swap">
+              <LandingSwap />
+            </Link>
+          </TraceEvent>
+        </LandingSwapContainer>
         <Gradient isDarkMode={isDarkMode} />
+        <GlowContainer>
+          <Glow />
+        </GlowContainer>
         <ContentContainer isDarkMode={isDarkMode}>
           <TitleText isDarkMode={isDarkMode}>Trade crypto & NFTs with confidence</TitleText>
           <SubTextContainer>
@@ -208,7 +308,32 @@ export default function Landing() {
               </ButtonCTA>
             </TraceEvent>
           </ActionsContainer>
+          <LearnMoreContainer
+            onClick={() => {
+              cardsRef?.current?.scrollIntoView({ behavior: 'smooth' })
+            }}
+          >
+            <Trans>Learn more</Trans>
+            <LearnMoreArrow />
+          </LearnMoreContainer>
         </ContentContainer>
+        <AboutContentContainer isDarkMode={isDarkMode}>
+          <CardGrid cols={2} ref={cardsRef}>
+            {MAIN_CARDS.map(({ darkBackgroundImgSrc, lightBackgroundImgSrc, ...card }) => (
+              <Card
+                {...card}
+                backgroundImgSrc={isDarkMode ? darkBackgroundImgSrc : lightBackgroundImgSrc}
+                key={card.title}
+              />
+            ))}
+          </CardGrid>
+          <CardGrid cols={3}>
+            {MORE_CARDS.map(({ darkIcon, lightIcon, ...card }) => (
+              <Card {...card} icon={isDarkMode ? darkIcon : lightIcon} key={card.title} type={CardType.Secondary} />
+            ))}
+          </CardGrid>
+          <ProtocolBanner />
+        </AboutContentContainer>
       </PageContainer>
     </Trace>
   )
